devenv/FontMgr: Split off gfx/nvn/DebugFontMgrNvn
